+++ pgbelt/__init__.py
@@ -1,3 +1,4 @@
 from pgbelt.config.models import DbConfig
 from pgbelt.config.models import DbupgradeConfig
 from pgbelt.config.remote import BaseResolver
+from pgbelt.config.remote import BaseSideResolver